+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Note
 
+- Use `TextAutoSize` to find the optimal text size
 - Remove predictive back opt-in after API 36
 
 ## [2.0.3](https://github.com/ldeso/blitz/releases/tag/v2.0.3) – 2025-04-20